requirements.txt
connexion[flask,swagger-ui,uvicorn]>=2.14.0
Flask>=2.0.0
PyYAML>=6.0
python-dateutil>=2.8.2
SQLAlchemy>=2.0.0  # If you'll be using a database
marshmallow>=3.15.0
